+++ src/test/java/com/meetup/server/global/clients/ratelimit/RateLimiterTest.java
@@ -1,4 +1,4 @@
-package com.meetup.server.global.clients.util;
+package com.meetup.server.global.clients.ratelimit;
 
 import com.meetup.server.support.IntegrationTestContainer;
 import org.junit.jupiter.api.Assertions;
